[not verified] Licensing: Move endpoints to own class
